% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{format.parser}
\alias{format.parser}
\title{Encode for printing}
\usage{
\method{format}{parser}(x, ...)
}
\arguments{
\item{x}{parser to be formatted}

\item{...}{other arguments to be passed to \code{format.character}}
}
\description{
Format a \code{parser} object for printing
}
\examples{
# No name, default formatter
format(parser('[0-9]+'))
# Custom name and formatter
format(parser('[0-9]+]', as.integer, name='int'))

}
