% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{formatter}
\alias{formatter}
\alias{formatter<-}
\title{Formatters}
\usage{
formatter(x)

formatter(x) <- value
}
\arguments{
\item{x}{parser}

\item{value}{formatter function to be set}
}
\value{
The formatter attribute (should be a function) for the passed object 
        (usually a \code{parser} object)
}
\description{
Get or set the formatter for a parser
}
\examples{
p <- parser('[0-9]+]')

# Default formatter
formatter(p)

# Set formatter
formatter(p) <- as.integer

# Custom formatter
formatter(p)

}
