% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{print.parser}
\alias{print.parser}
\title{Print}
\usage{
\method{print}{parser}(x, ...)
}
\arguments{
\item{x}{parser to be printed}

\item{...}{Other arguments; ignored}
}
\value{
\code{x}, invisibly
}
\description{
Print a \code{parser} object. Underlying method uses \code{cat}.
}
\examples{
# No name, default formatter
print(parser('[0-9]+'))

#Custom name and formatter
print(parser('[0-9]+]', as.integer, name='int'))

}
